requirements.txt
fastapi
uvicorn
langchain
openai
langchain-community
langchain-core
langchain-openai
python-decouple
tavily-python
openpyxl

